refactored, added 2 additional classes
