+++ lab1/src/Main.java
@@ -4,13 +4,13 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class CSVWorker {
+public class Main {
     public static void main(String[] args) {
         try(BufferedReader reader  = new BufferedReader(new FileReader("input.txt"));
             BufferedWriter writer = new BufferedWriter(new FileWriter("output.csv")))
         {
-            WordClass arr = new WordClass();
-            arr.printDataInCSV(reader, writer);
+            CsvFormatPrinter csvPrinter = new CsvFormatPrinter();
+            csvPrinter.printDataInCSV(reader, writer);
         } catch (IOException ex) {
             System.err.println("Error while reading file:" + ex.getLocalizedMessage());
         }
